+++ src/eval.c
@@ -292,3 +292,22 @@ bool eval_possible(eval_t *e, tree_t t)
       return eval_not_possible(e, t, tree_kind_str(tree_kind(t)));
    }
 }
+
+tree_t eval_case(eval_t *ex, tree_t stmt)
+{
+   assert(tree_kind(stmt) == T_CASE_GENERATE);
+
+   vcode_unit_t thunk = lower_thunk(stmt);
+
+   jit_set_silent(ex->jit, false);
+   jit_limit_backedges(ex->jit, 0);
+
+   jit_scalar_t result = { .integer = -1 };
+   if (!jit_call_thunk(ex->jit, thunk, &result))
+      error_at(tree_loc(tree_value(stmt)), "generate expression is not static");
+
+   if (result.integer == -1)
+      return NULL;
+   else
+      return tree_stmt(stmt, result.integer);
+}