+++ NEWS.md
@@ -22,6 +22,7 @@
 - Negation of the smallest negative value of a type such as
   `-integer'left` now produces an error.
 - Default OSVVM version updated to 2022.11.
+- `case .. generate` statements are now supported in VHDL-2008.
 
 ## Version 1.7.2 - 2022-10-16
 - Fixed build on FreeBSD/arm (#534).